test(fix): add helper class for dayago and move navigation S4 tests in a package
