+++ app/src/androidTest/java/com/android/mygarden/ui/navigation/navS4/NavigationS4TestsEditProfile.kt
@@ -1,4 +1,4 @@
-package com.android.mygarden.ui.navigation
+package com.android.mygarden.ui.navigation.navS4
 
 import androidx.compose.ui.test.assertIsDisplayed
 import androidx.compose.ui.test.assertTextContains
@@ -11,6 +11,8 @@ import androidx.navigation.compose.rememberNavController
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.android.mygarden.model.profile.ProfileRepositoryProvider
 import com.android.mygarden.ui.garden.GardenScreenTestTags
+import com.android.mygarden.ui.navigation.AppNavHost
+import com.android.mygarden.ui.navigation.Screen
 import com.android.mygarden.ui.profile.ProfileScreenTestTags
 import com.android.mygarden.ui.theme.MyGardenTheme
 import com.android.mygarden.utils.FakeProfileRepository
